Remove deprecated coroutines test bridges
Move dispatcher test helpers to the junit5 owner module and migrate Flow assertion call sites to the assertions owner module so the coroutines module no longer publishes deprecated bridge helpers.

Constraint: Issue #879 requires no active io.bluetape4k.coroutines.tests imports outside historical docs.

Rejected: Keeping compatibility bridges in bluetape4k-coroutines | the issue explicitly removes the deprecated bridge package.

Rejected: Adding a core dependency to bluetape4k-junit5 | helper validation and cleanup are simple enough to keep local.

Confidence: high

Scope-risk: moderate

Directive: Future Flow assertion tests should import io.bluetape4k.assertions.coroutines directly; dispatcher helper tests should import io.bluetape4k.junit5.coroutines.

Tested: ./gradlew :bluetape4k-junit5:compileKotlin

Tested: ./gradlew :bluetape4k-junit5:test :bluetape4k-assertions:test

Tested: ./gradlew :bluetape4k-coroutines:compileTestKotlin :bluetape4k-coroutines:test

Tested: ./gradlew :bluetape4k-examples-coroutines-demo:compileTestKotlin

Tested: rg "io\.bluetape4k\.coroutines\.tests" -g '*.kt' -g '*.kts'

Tested: git diff --check

Not-tested: Full repository build; scope is targeted to coroutines, assertions, junit5, and coroutines demo compile.
