+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/FlatMapDropTest.kt
@@ -1,6 +1,6 @@
 package io.bluetape4k.coroutines.flow.extensions
 
-import io.bluetape4k.coroutines.tests.assertResult
+import io.bluetape4k.assertions.coroutines.assertResult
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.onEach
