+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/MergeFlowsTest.kt
@@ -1,7 +1,7 @@
 package io.bluetape4k.coroutines.flow.extensions
 
-import io.bluetape4k.coroutines.tests.assertResult
-import io.bluetape4k.coroutines.tests.assertResultSet
+import io.bluetape4k.assertions.coroutines.assertResult
+import io.bluetape4k.assertions.coroutines.assertResultSet
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
