+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/RaceTest.kt
@@ -1,7 +1,7 @@
 package io.bluetape4k.coroutines.flow.extensions
 
 import app.cash.turbine.test
-import io.bluetape4k.coroutines.tests.assertResult
+import io.bluetape4k.assertions.coroutines.assertResult
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.Flow
