+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/ZipWithNextTest.kt
@@ -1,7 +1,7 @@
 package io.bluetape4k.coroutines.flow.extensions
 
 import app.cash.turbine.test
-import io.bluetape4k.coroutines.tests.assertError
+import io.bluetape4k.assertions.coroutines.assertError
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import kotlinx.coroutines.flow.emptyFlow
 import kotlinx.coroutines.flow.flow
