+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/parallel/ParallelFlowConcatMapTest.kt
@@ -2,11 +2,11 @@ package io.bluetape4k.coroutines.flow.extensions.parallel
 
 import io.bluetape4k.coroutines.flow.extensions.flowRangeOf
 import io.bluetape4k.coroutines.flow.extensions.log
-import io.bluetape4k.coroutines.tests.assertError
-import io.bluetape4k.coroutines.tests.assertFailure
-import io.bluetape4k.coroutines.tests.assertResult
-import io.bluetape4k.coroutines.tests.assertResultSet
-import io.bluetape4k.coroutines.tests.withParallels
+import io.bluetape4k.assertions.coroutines.assertError
+import io.bluetape4k.assertions.coroutines.assertFailure
+import io.bluetape4k.assertions.coroutines.assertResult
+import io.bluetape4k.assertions.coroutines.assertResultSet
+import io.bluetape4k.junit5.coroutines.withParallels
 import io.bluetape4k.junit5.coroutines.runSuspendTest
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import io.bluetape4k.logging.trace
