+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/parallel/ParallelFlowReduceTest.kt
@@ -2,8 +2,8 @@ package io.bluetape4k.coroutines.flow.extensions.parallel
 
 import io.bluetape4k.coroutines.flow.extensions.flowRangeOf
 import io.bluetape4k.coroutines.flow.extensions.log
-import io.bluetape4k.coroutines.tests.assertResult
-import io.bluetape4k.coroutines.tests.withParallels
+import io.bluetape4k.assertions.coroutines.assertResult
+import io.bluetape4k.junit5.coroutines.withParallels
 import io.bluetape4k.junit5.coroutines.runSuspendTest
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import io.bluetape4k.logging.trace
