+++ bluetape4k/coroutines/src/test/kotlin/io/bluetape4k/coroutines/flow/extensions/subject/ReplaySubjectSizeAndTimeBoundTest.kt
@@ -2,7 +2,7 @@ package io.bluetape4k.coroutines.flow.extensions.subject
 
 import io.bluetape4k.coroutines.flow.extensions.log
 import io.bluetape4k.coroutines.support.log
-import io.bluetape4k.coroutines.tests.withSingleThread
+import io.bluetape4k.junit5.coroutines.withSingleThread
 import io.bluetape4k.junit5.coroutines.runSuspendTest
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import kotlinx.coroutines.cancel
