+++ examples/coroutines-demo/src/test/kotlin/io/bluetape4k/examples/coroutines/flow/FlowBuilderExamples.kt
@@ -2,7 +2,7 @@ package io.bluetape4k.examples.coroutines.flow
 
 import app.cash.turbine.test
 import io.bluetape4k.coroutines.flow.extensions.log
-import io.bluetape4k.coroutines.tests.assertResult
+import io.bluetape4k.assertions.coroutines.assertResult
 import io.bluetape4k.logging.coroutines.KLoggingChannel
 import io.bluetape4k.logging.debug
 import kotlinx.coroutines.coroutineScope
